Implement private persistent rooms

+++ app/server/skychat/plugins/core/RoomManagerPlugin.ts
@@ -2,18 +2,24 @@ import {Connection} from "../../Connection";
 import {RoomPlugin} from "../../RoomPlugin";
 import { UserController } from "../../UserController";
 import { Config } from "../../Config";
+import { Room } from "../../Room";
+import { Session } from "../../Session";
 
 
 export class RoomManagerPlugin extends RoomPlugin {
 
     static readonly commandName = 'room';
 
-    static readonly commandAliases = ['roomset', 'roomcreate', 'roomdelete', 'roomplugin']
-
-    readonly opOnly = true;
+    static readonly commandAliases = ['roomset', 'roomcreate', 'roomleave', 'roomdelete', 'roomplugin'];
 
     readonly rules = {
         room: { },
+        roomcreate: {
+            minCount: 0,
+            params: [
+                {name: 'name', pattern: /.+/},
+            ]
+        },
         roomset: {
             minCount: 2,
             params: [
@@ -21,14 +27,8 @@ export class RoomManagerPlugin extends RoomPlugin {
                 {name: 'value', pattern: /.?/},
             ]
         },
-        roomcreate: {
-            minCount: 2,
-            params: [
-                {name: 'new room id', pattern: /^([0-9]+)$/},
-                {name: 'name', pattern: /.+/},
-            ]
-        },
-        roomdelete: {maxCount: 0,},
+        roomleave: { minCount: 0, },
+        roomdelete: { maxCount: 0, },
         roomplugin: {
             minCount: 1,
             maxCount: 1,
@@ -50,6 +50,10 @@ export class RoomManagerPlugin extends RoomPlugin {
                 await this.handleRoomCreate(param, connection);
                 break;
 
+            case 'roomleave':
+                await this.handleRoomLeave(param, connection);
+                break;
+
             case 'roomdelete':
                 await this.handleRoomDelete(param, connection);
                 break;
@@ -60,7 +64,28 @@ export class RoomManagerPlugin extends RoomPlugin {
         }
     }
 
+    async handleRoomCreate(param: string, connection: Connection): Promise<void> {
+        if (! connection.session.isOP()) {
+            throw new Error('Only OP can create public rooms');
+        }
+        const roomName = param.trim();
+        const room = this.room.manager.createRoom(roomName);
+        const message = UserController.createNeutralMessage({ id: 0, content: `Room ${room.id} has been created` });
+        connection.send('message', message.sanitized());
+    }
+
+    canManageRoom(session: Session, room: Room): boolean {
+        if (room.isPrivate) {
+            return room.whitelist.indexOf(session.identifier) !== -1;
+        } else {
+            return session.isOP();
+        }
+    }
+
     async handleRoomSet(param: string, connection: Connection): Promise<void> {
+        if (! this.canManageRoom(connection.session, this.room)) {
+            throw new Error('You do not have the permission to modify this room');
+        }
         const property = param.substr(0, param.indexOf(' '));
         const value = param.substr(param.indexOf(' ') + 1).trim();
         
@@ -73,28 +98,41 @@ export class RoomManagerPlugin extends RoomPlugin {
             default:
                 throw new Error(`Invalid property ${property}`);
         }
+        Object.values(Session.sessions).forEach(session => this.room.manager.sendRoomList(session));
         const message = UserController.createNeutralMessage({ id: 0, content: `Room property ${property} set to ${value}` });
         connection.send('message', message.sanitized());
     }
 
-    async handleRoomCreate(param: string, connection: Connection): Promise<void> {
-        const roomId = parseInt(param.substr(0, param.indexOf(' ')));
-        const roomName = param.substr(param.indexOf(' ') + 1).trim();
-        if (isNaN(roomId)) {
-            throw new Error('Invalid room id');
+    async handleRoomLeave(param: string, connection: Connection): Promise<void> {
+        if (! this.room.isPrivate) {
+            throw new Error('You can not leave a public room');
         }
-        const room = this.room.manager.createRoom(roomId, roomName);
-        const message = UserController.createNeutralMessage({ id: 0, content: `Room ${room.id} has been created` });
-        connection.send('message', message.sanitized());
+        if (this.room.whitelist.length === 1) {
+            throw new Error('You can not leave this room, you are the last user in it');
+        }
+        this.room.unallow(connection.session.identifier);
+        this.room.manager.sendRoomList(connection);
     }
 
     async handleRoomDelete(param: string, connection: Connection): Promise<void> {
+        if (this.room.isPrivate && this.room.whitelist.length > 1) {
+            throw new Error('You can not delete this room, others are still in it');
+        }
+        if (! this.canManageRoom(connection.session, this.room)) {
+            throw new Error('You do not have the permission to delete this room');
+        }
         await this.room.manager.deleteRoom(this.room.id);
         const message = UserController.createNeutralMessage({ id: 0, content: `Room ${this.room.id} has been deleted` });
         connection.send('message', message.sanitized());
     }
 
     async handleRoomPlugin(param: string, connection: Connection): Promise<void> {
+        if (this.room.isPrivate) {
+            throw new Error('Unable to customize plugins in private rooms');
+        }
+        if (! connection.session.isOP()) {
+            throw new Error('Only OP can manage public room plugins');
+        }
         const add = param[0] === '+';
         const pluginName = param.substr(1);
